Age & Gender Confirmation for Aptitude Test
I am still learning but I was able to create a java program that restricts users of a certain age and confirms their gender for an aptitude test, this can be used for betting as well, it also greets the user based on if you are male or female and then proceeds to ask you questions, and scores you, I need someone to update this to tell the person his score e.g. Mr. Kayode your score is ???; this should be based on input of name.